requirements.txt
flask==3.0.0
werkzeug==3.0.1
beautifulsoup4==4.9.3
requests==2.26.0
openai==0.27.0
python-dotenv==0.19.0
playwright==1.41.2
asgiref==3.8.1
gunicorn==21.2.0
gevent==23.9.1
supervisor==4.2.5
sentry-sdk[flask]==1.39.1
prometheus-client==0.19.0
flask-limiter==3.5.0
